User data fetching moved to higher order component(page) for addTodo component
